WL-1125 Refactored retrieval of results to handle .auth submission as desired.
This just handles the list of actual submissions rather than cross-referencing
this against the list of potential submitters. Potential submitters without a
submission were previously assigned a 'phantom' submission, but this seemed
redundant. The lists of userIds and users also 'fell out' as being redundant
too. The only sticking point at the moment is that the security check to verify
that each submission is from someone who can still submit may be computationally
expensive (?)