WL-1125 Show .auth users in the results
WL-1125 Refactored retrieval of results to handle .auth submission as desired.

This just handles the list of actual submissions rather than cross-referencing
this against the list of potential submitters. Potential submitters without a
submission were previously assigned a 'phantom' submission, but this seemed
redundant. The lists of userIds and users also 'fell out' as being redundant
too. The only sticking point at the moment is that the security check to verify
that each submission is from someone who can still submit may be computationally
expensive (?)

WL-1125 Changed implementation to return submissions in all required cases.

The method in question needs to cater for an orthogonal set of cases; a known
group (with phantom submissions assigned to users who are missing one) and
open-ended groups such as .auth, where only the submissions that are actually
found are of interest. Unfortunately I made erroneous assumptions with my
previous commit regarding redundant variables; they are required after all! As
before, there is a possible performance concern around checking that each
submission is from a user who can still submit. However, the intent is to deal
with this *if* and when it arises.