WL-1125 Changed implementation to return submissions in all required cases.
The method in question needs to cater for an orthogonal set of cases; a known
group (with phantom submissions assigned to users who are missing one) and
open-ended groups such as .auth, where only the submissions that are actually
found are of interest. Unfortunately I made erroneous assumptions with my
previous commit regarding redundant variables; they are required after all! As
before, there is a possible performance concern around checking that each
submission is from a user who can still submit. However, the intent is to deal
with this *if* and when it arises.