WL-1158 Added a boolean property to assessment to send an e-mail / certicate.
This is a boolean property to control whether or not an e-mail gets sent when
a user submits their assessment (a notion of a 'pass mark' is coming later).
This is currently the minimum required (it omits Oracle support and does not
include conditional support for surveys where the concept makes less sense).
However, it goes from UI, through Java Beans / DAO and down to the persistence
layer. The e-mail sending itself does not currently make use of this yet.