WL-1158 Removed redundant methods around the sendEmail property.
Several of the properties in the AssessmentImpl class have associated initXxx()
and getXxxChanged() methods. I originally implemented these for
sendEmailOnSubmission because I thought they were the reason my code was failing
(it actually turned out to be my failure to modify the 'clone' method!). On
closer inspection, it seems as though the properties that have these methods
implemented are ones whose (change of) value has an impact on the integration
with GradeBook; as neither sendEmailOnSubmission nor passMark do, I have chosen
not to implement them for these in order to maximise maintainability and
legibility.