WL-1158 Added a float property to assessment to capture a pass mark.
This is an optional float property intended to be used in conjunction with
sendEmailOnSubmission, to control whether or not an e-mail gets sent when a user
submits their assessment. It used to specify the miniumum percentage score
required before an e-mail is sent. This is currently the minimum required (it
omits Oracle support and does not include conditional support for surveys where
the concept makes less sense). However, it goes from UI, through Java Beans /
DAO and down to the persistence layer. The UI for setting this is currently
simplistic, but I intend to make it more consistent with other analogous mneme
widgets. (Note: The e-mail sending itself does not currently make use of this
yet).