WL-1264 Added an 'OSS ID' column to the class list of student marks.
(This screen is reached via Marking > Mark Submissions). The solution may look
rather heavyweight, but there is currently no way of using a key to reference a
value in a map from within an ambrosia template. To have added the equivalent to
a pre-existing CSV export would have undoubtedly been much simpler!