WL-1158 Made tweaks to the UI widgets that set the pass mark and send e-mails.
Part of the the intent here was to capture the fact that a pass mark in itself
is quite orthogonal to e-mail sending and potentially more fundamental, so I
moved it up the list of controls (and changed the wording of the labels
accordingly). I also made the pass mark box narrower (6 columns equates to
'100.00').